test: increase coverage docs: simplify
