+++ src/__mocks__/purge.js
@@ -1,0 +1,5 @@
+import { jest } from '@jest/globals'
+
+const purge = jest.fn()
+
+export { purge }
