+++ src/loader.js
@@ -1,4 +1,4 @@
-import { purge } from './purge'
+import { purge } from './purge.js'
 
 const loader = function (source, map, meta) {
   this.callback(null, purge(source), map, meta)
